fix(arborist): only replace hostname for resolved URL (#8185)
Related to #6110 

There is an edge case where if an npm project is using a registry with a
path the `replace-registry-host` is including the path from the registry
url.

Current Behavior: 

- Inputs: 
- `registry=https://internal.mycompany.com/artifactory/api/npm/npm-all`
   - `replace-registry-host=always`
- resolved: "https://external.mycompany.com/artifactory/api/npm/npm-all"
(package-lock.json)

- Output:
"https://internal.mycompany.com/artifactory/api/npm/npm-all/api/npm/npm-all"

Expected Behavior:
- Inputs: 
- `registry=https://internal.mycompany.com/artifactory/api/npm/npm-all`
   - `replace-registry-host=always`
- resolved: "https://external.mycompany.com/artifactory/api/npm/npm-all"
(package-lock.json)

- Output: "https://internal.mycompany.com/artifactory/api/npm/npm-all"

This fix resolves this inconsistency by refactoring the logic to
construct URL objects from the registry and resolved strings and maps
the hostname of the new registry to the hostname of the resolved url
instead of doing string parsing.

I'm welcome to any feedback on this solution!